requirements.txt
asgiref==3.3.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.0
django-filter==2.4.0
django-jazzmin==2.4.7
django-mysql==3.11.1
django-phone-field==1.8.1
django-templated-mail==1.1.1
djangorestframework==3.10.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==2.10
importlib-metadata==1.7.0
itypes==1.2.0
Jinja2==3.0.1
Markdown==3.3.4
MarkupSafe==2.0.1
mysqlclient==2.0.3
notes==0.3.0
oauthlib==3.1.1
Pillow==8.1.2
psycopg2==2.9.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.4.1

